Remove the old gradient test suite, and replace by the modern one
The "gradients" test suite was built before gradients reach their final
syntax, and only checked them with a prefixed syntax. The
"unprefixed-gradients" is a rework of it, updating to the final syntax
as specified per the w3c, and supports the unprefixed syntax. Opera
needed to keep both for historical reasons, but only the later is
relevant to the rest of the world.